WIP - refactored parse xml script, renamed and reorganized other scripts
